+++ iddd_identityaccess/Domain.Model/Identity/ContactInformation.cs
@@ -103,153 +103,4 @@ namespace SaaSOvation.IdentityAccess.Domain.Model.Identity
             yield return this.SecondaryTelephone;
         }
     }
-
-    public class EmailAddress : ValueObject
-    {
-        public EmailAddress(String address)
-        {
-            this.Address = address;
-        }
-
-        public EmailAddress(EmailAddress emailAddress)
-            : this(emailAddress.Address)
-        {
-        }
-
-        internal EmailAddress()
-        {
-        }
-
-        string _address;
-        
-        public string Address
-        {
-            get
-            {
-                return this._address;
-            }
-            set
-            {
-                AssertionConcern.AssertArgumentNotEmpty(value, "The email address is required.");
-                AssertionConcern.AssertArgumentLength(value, 1, 100, "Email address must be 100 characters or less.");
-                AssertionConcern.AssertArgumentMatches(
-                        "\\w+([-+.']\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*",
-                        value,
-                        "Email address format is invalid.");
-
-                this._address = value;
-            }
-        }
-
-        public override string ToString()
-        {
-            return "EmailAddress [address=" + Address + "]";
-        }
-
-        protected override System.Collections.Generic.IEnumerable<object> GetEqualityComponents()
-        {
-            yield return _address.ToUpper();
-        }
-    }
-
-    public class PostalAddress : ValueObject
-    {
-        public PostalAddress(
-                String streetAddress,
-                String city,
-                String stateProvince,
-                String postalCode,
-                String countryCode)
-        {
-            this.City = city;
-            this.CountryCode = countryCode;
-            this.PostalCode = postalCode;
-            this.StateProvince = stateProvince;
-            this.StreetAddress = streetAddress;
-        }
-
-        public PostalAddress(PostalAddress postalAddress)
-            : this(postalAddress.StreetAddress,
-                   postalAddress.City,
-                   postalAddress.StateProvince,
-                   postalAddress.PostalCode,
-                   postalAddress.CountryCode)
-        {
-        }
-
-        public string City { get; private set; }
-
-        public string CountryCode { get; private set; }
-
-        public string PostalCode { get; private set; }
-
-        public string StateProvince { get; private set; }
-
-        public string StreetAddress { get; private set; }
-
-        public override string ToString()
-        {
-            return "PostalAddress [streetAddress=" + StreetAddress
-                    + ", city=" + City + ", stateProvince=" + StateProvince
-                    + ", postalCode=" + PostalCode
-                    + ", countryCode=" + CountryCode + "]";
-        }
-
-        protected override System.Collections.Generic.IEnumerable<object> GetEqualityComponents()
-        {
-            yield return this.StreetAddress;
-            yield return this.City;
-            yield return this.StateProvince;
-            yield return this.PostalCode;
-            yield return this.CountryCode;
-        }
-    }
-
-    public class Telephone : ValueObject
-    {
-        public Telephone(String number)
-        {
-            this.Number = number;
-        }
-
-        public Telephone(Telephone telephone)
-            : this(telephone.Number)
-        {
-        }
-
-        internal Telephone()
-        {
-        }
-
-        string _number;
-        public string Number
-        {
-            get
-            {
-                return this._number;
-            }
-
-            set
-            {
-                AssertionConcern.AssertArgumentNotEmpty(value, "Telephone number is required.");
-                AssertionConcern.AssertArgumentLength(value, 5, 20, "Telephone number may not be more than 20 characters.");
-                AssertionConcern.AssertArgumentMatches(
-                        "((\\(\\d{3}\\))|(\\d{3}-))\\d{3}-\\d{4}",
-                        value,
-                        "Telephone number or its format is invalid.");
-
-                this._number = value;
-            }
-        }
-
-        public override string ToString()
-        {
-            return "Telephone [number=" + Number + "]";
-        }
-
-        protected override System.Collections.Generic.IEnumerable<object> GetEqualityComponents()
-        {
-            yield return this.Number;
-        }
-    }
 }
