--- iddd_identityaccess/Domain.Model/Identity/GroupEvents.cs
@@ -1,107 +1,0 @@
-﻿// Copyright 2012,2013 Vaughn Vernon
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-namespace SaaSOvation.IdentityAccess.Domain.Model.Identity
-{
-    using System;
-    using SaaSOvation.Common.Domain.Model;
-
-    public class GroupGroupAdded : IDomainEvent
-    {
-        public GroupGroupAdded(TenantId tenantId, string groupName, string nestedGroupName)
-        {
-            this.EventVersion = 1;
-            this.GroupName = groupName;
-            this.NestedGroupName = nestedGroupName;
-            this.OccurredOn = new DateTime();
-            this.TenantId = tenantId.Id;
-        }
-
-        public int EventVersion { get; set; }
-
-        public string GroupName { get; private set; }
-
-        public string NestedGroupName { get; private set; }
-
-        public DateTime OccurredOn { get; set; }
-
-        public string TenantId { get; private set; }
-    }
-
-    public class GroupGroupRemoved : IDomainEvent
-    {
-        public GroupGroupRemoved(TenantId tenantId, string groupName, string nestedGroupName)
-        {
-            this.EventVersion = 1;
-            this.GroupName = groupName;
-            this.NestedGroupName = nestedGroupName;
-            this.OccurredOn = new DateTime();
-            this.TenantId = tenantId.Id;
-        }
-
-        public int EventVersion { get; set; }
-
-        public string GroupName { get; private set; }
-
-        public string NestedGroupName { get; private set; }
-
-        public DateTime OccurredOn { get; set; }
-
-        public string TenantId { get; private set; }
-    }
-
-    public class GroupUserAdded : IDomainEvent
-    {
-        public GroupUserAdded(TenantId tenantId, string groupName, string username)
-        {
-            this.EventVersion = 1;
-            this.GroupName = groupName;
-            this.OccurredOn = new DateTime();
-            this.TenantId = tenantId.Id;
-            this.Username = username;
-        }
-
-        public int EventVersion { get; set; }
-
-        public string GroupName { get; private set; }
-
-        public DateTime OccurredOn { get; set; }
-
-        public string TenantId { get; private set; }
-
-        public string Username { get; private set; }
-    }
-
-    public class GroupUserRemoved : IDomainEvent
-    {
-        public GroupUserRemoved(TenantId tenantId, string groupName, string username)
-        {
-            this.EventVersion = 1;
-            this.GroupName = groupName;
-            this.OccurredOn = new DateTime();
-            this.TenantId = tenantId.Id;
-            this.Username = username;
-        }
-
-        public int EventVersion { get; set; }
-
-        public string GroupName { get; private set; }
-
-        public DateTime OccurredOn { get; set; }
-
-        public string TenantId { get; private set; }
-
-        public string Username { get; private set; }
-    }
-}
