+++ iddd_identityaccess/Domain.Model/Identity/GroupMemberType.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace SaaSOvation.IdentityAccess.Domain.Model.Identity
+{
+    public enum GroupMemberType 
+    { 
+        Group, 
+        User 
+    }
+}
