+++ iddd_identityaccess/Domain.Model/Identity/GroupProvisioned.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace SaaSOvation.IdentityAccess.Domain.Model.Identity
+{
+    public class GroupProvisioned : SaaSOvation.Common.Domain.Model.IDomainEvent
+    {
+        public GroupProvisioned(TenantId tenantId, string name)
+        {
+            this.EventVersion = 1;
+            this.Name = name;
+            this.OccurredOn = DateTime.Now;
+            this.TenantId = tenantId.Id;
+        }
+
+        public int EventVersion { get; set; }
+
+        public string Name { get; private set; }
+
+        public DateTime OccurredOn { get; set; }
+
+        public string TenantId { get; private set; }
+    }
+}
