+++ iddd_identityaccess/Domain.Model/Identity/PersonNameChanged.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace SaaSOvation.IdentityAccess.Domain.Model.Identity
+{
+    public class PersonNameChanged : SaaSOvation.Common.Domain.Model.IDomainEvent
+    {
+        public PersonNameChanged(
+                TenantId tenantId,
+                String username,
+                FullName name)
+        {
+            this.EventVersion = 1;
+            this.Name = name;
+            this.OccurredOn = DateTime.Now;
+            this.TenantId = tenantId.Id;
+            this.Username = username;
+        }
+
+        public int EventVersion { get; set; }
+
+        public FullName Name { get; private set; }
+
+        public DateTime OccurredOn { get; set; }
+
+        public string TenantId { get; private set; }
+
+        public string Username { get; private set; }
+    }
+}
