+++ iddd_identityaccess/Domain.Model/Identity/UserPasswordChanged.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace SaaSOvation.IdentityAccess.Domain.Model.Identity
+{
+    public class UserPasswordChanged : SaaSOvation.Common.Domain.Model.IDomainEvent
+    {
+        public UserPasswordChanged(
+                TenantId tenantId,
+                String username)
+        {
+            this.EventVersion = 1;
+            this.OccurredOn = DateTime.Now;
+            this.TenantId = tenantId.Id;
+            this.Username = username;
+        }
+
+        public int EventVersion { get; set; }
+
+        public DateTime OccurredOn { get; set; }
+
+        public string TenantId { get; private set; }
+
+        public string Username { get; private set; }
+    }
+}
