+++ iddd_identityaccess/Domain.Model/Identity/UserRegistered.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace SaaSOvation.IdentityAccess.Domain.Model.Identity
+{
+    public class UserRegistered : SaaSOvation.Common.Domain.Model.IDomainEvent
+    {
+        public UserRegistered(
+                TenantId tenantId,
+                String username,
+                FullName name,
+                EmailAddress emailAddress)
+        {
+            this.EmailAddress = emailAddress;
+            this.EventVersion = 1;
+            this.Name = name;
+            this.OccurredOn = DateTime.Now;
+            this.TenantId = tenantId.Id;
+            this.Username = username;
+        }
+
+        public EmailAddress EmailAddress { get; private set; }
+
+        public int EventVersion { get; set; }
+
+        public FullName Name { get; private set; }
+
+        public DateTime OccurredOn { get; set; }
+
+        public string TenantId { get; private set; }
+
+        public string Username { get; private set; }
+    }
+}
