+++ iddd_identityaccess/IdentityAccess.csproj
@@ -66,32 +66,51 @@
     <Compile Include="Application\IdentityApplicationService.cs" />
     <Compile Include="Application\NotificationApplicationService.cs" />
     <Compile Include="Domain.Model\Access\AuthorizationService.cs" />
+    <Compile Include="Domain.Model\Access\GroupUnassignedFromRole.cs" />
     <Compile Include="Domain.Model\Access\Role.cs" />
-    <Compile Include="Domain.Model\Access\RoleEvents.cs" />
+    <Compile Include="Domain.Model\Access\GroupAssignedToRole.cs" />
     <Compile Include="Domain.Model\Access\IRoleRepository.cs" />
+    <Compile Include="Domain.Model\Access\RoleProvisioned.cs" />
+    <Compile Include="Domain.Model\Access\UserAssginedToRole.cs" />
+    <Compile Include="Domain.Model\Access\UserUnassignedFromRole.cs" />
     <Compile Include="Domain.Model\Identity\AuthenticationService.cs" />
     <Compile Include="Domain.Model\Identity\ContactInformation.cs" />
+    <Compile Include="Domain.Model\Identity\EmailAddress.cs" />
     <Compile Include="Domain.Model\Identity\Enablement.cs" />
+    <Compile Include="Domain.Model\Identity\GroupGroupRemoved.cs" />
+    <Compile Include="Domain.Model\Identity\GroupMemberType.cs" />
+    <Compile Include="Domain.Model\Identity\GroupProvisioned.cs" />
+    <Compile Include="Domain.Model\Identity\GroupUserAdded.cs" />
+    <Compile Include="Domain.Model\Identity\GroupUserRemoved.cs" />
     <Compile Include="Domain.Model\Identity\IEncryptionService.cs" />
     <Compile Include="Domain.Model\Identity\FullName.cs" />
     <Compile Include="Domain.Model\Identity\Group.cs" />
-    <Compile Include="Domain.Model\Identity\GroupEvents.cs" />
+    <Compile Include="Domain.Model\Identity\GroupGroupAdded.cs" />
     <Compile Include="Domain.Model\Identity\GroupMember.cs" />
     <Compile Include="Domain.Model\Identity\GroupMemberService.cs" />
     <Compile Include="Domain.Model\Identity\IGroupRepository.cs" />
     <Compile Include="Domain.Model\Identity\InvitationDescriptor.cs" />
     <Compile Include="Domain.Model\Identity\PasswordService.cs" />
     <Compile Include="Domain.Model\Identity\Person.cs" />
+    <Compile Include="Domain.Model\Identity\PersonNameChanged.cs" />
+    <Compile Include="Domain.Model\Identity\PostalAddress.cs" />
     <Compile Include="Domain.Model\Identity\RegistrationInvitation.cs" />
+    <Compile Include="Domain.Model\Identity\Telephone.cs" />
     <Compile Include="Domain.Model\Identity\Tenant.cs" />
-    <Compile Include="Domain.Model\Identity\TenantEvents.cs" />
+    <Compile Include="Domain.Model\Identity\TenantActivated.cs" />
+    <Compile Include="Domain.Model\Identity\TenantDeactivated.cs" />
+    <Compile Include="Domain.Model\Identity\TenantAdministratorRegistered.cs" />
     <Compile Include="Domain.Model\Identity\TenantId.cs" />
+    <Compile Include="Domain.Model\Identity\TenantProvisioned.cs" />
     <Compile Include="Domain.Model\Identity\TenantProvisioningService.cs" />
     <Compile Include="Domain.Model\Identity\ITenantRepository.cs" />
     <Compile Include="Domain.Model\Identity\User.cs" />
     <Compile Include="Domain.Model\Identity\UserDescriptor.cs" />
-    <Compile Include="Domain.Model\Identity\UserEvents.cs" />
+    <Compile Include="Domain.Model\Identity\UserEnablementChanged.cs" />
+    <Compile Include="Domain.Model\Identity\PersonContactInformationChanged.cs" />
     <Compile Include="Domain.Model\Identity\IUserRepository.cs" />
+    <Compile Include="Domain.Model\Identity\UserPasswordChanged.cs" />
+    <Compile Include="Domain.Model\Identity\UserRegistered.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Domain.Model\DomainRegistry.cs" />
     <Compile Include="Infrastructure\Services\MD5EncryptionService.cs" />
@@ -102,9 +121,7 @@
       <Name>Common</Name>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Infrastructure\Persistence\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
